+++ src/DeepWoods/Map/DeepWoods.cs
@@ -6,7 +6,6 @@ using DeepWoodsMod.Stuff;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Netcode;
-using Newtonsoft.Json.Linq;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.BellsAndWhistles;