+++ src/DeepWoods/Stuff/HealingFountain.cs
@@ -3,7 +3,6 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
 using StardewValley.TerrainFeatures;
-using System;
 
 namespace DeepWoodsMod
 {